Try Fixing Disposal Issues with ExtractToDirectory(filename) API
This API was relying on the FileEntry finalizer to dispose the stream, but the timing of disposal is unreliable and users expect locks to be released after the method returns. This changes the filename based ExtractToDirectory API to create a stream in the method context and dispose it instead of relying on the FileEntry finalizer.

Fix #149.