Improve Dispose Behavior on ExtractToDirectory(filename) API (#150)
* Try Fixing Disposal Issues with ExtractToDirectory(filename) API

This API was relying on the FileEntry finalizer to dispose the stream, but the timing of disposal is unreliable and users expect locks to be released after the method returns. This changes the filename based ExtractToDirectory API to create a stream in the method context and dispose it instead of relying on the FileEntry finalizer.

Fix #149.

* Update Dependencies

* Add new tests to confirm lock has been released

Adds new test cases for the lock behavior on the ExtractToDirectory(filename) api.